refactor: 결과계산로직 분리
