+++ music/src/Pages/SecondTest/SecondPage9.js
@@ -1,62 +1,14 @@
 import React from 'react';
-import { useRecoilValue } from 'recoil';
-
-import RbBox from '../../Components/CheckBox/RbBox';
+import { useFavoriteGenre } from '../../hooks/useFavoriteGenre';
 import { GAUGE_PERCENTAGES } from '../../constants/gaugePercentages';
-import { PopValueState } from '../../Components/CheckBox/PopBox';
-import { HipValueState } from '../../Components/CheckBox/HipBox';
-import { JazzValueState } from '../../Components/CheckBox/JazzBox';
-import { RbValueState } from '../../Components/CheckBox/RbBox';
-import { RockValueState } from '../../Components/CheckBox/RockBox';
-
+import RbBox from '../../Components/CheckBox/RbBox';
 import Bar from '../../Components/common/Progress/Bar';
 import SecondBtn from '../../Components/common/Button/SecondBtn';
 import ScrollToNext from '../../Components/ScrollToNext/scrollToNext';
-
 import './SecondPage.css';
 
 const SecondPage9 = () => {
-  const PopCheckValue = useRecoilValue(PopValueState);
-  const HipCheckValue = useRecoilValue(HipValueState);
-  const JazzCheckValue = useRecoilValue(JazzValueState);
-  const RbCheckValue = useRecoilValue(RbValueState);
-  const RockCheckValue = useRecoilValue(RockValueState);
-
-  const PopTotal = Object.values(PopCheckValue).reduce(
-    (sum, value) => sum + value,
-    0,
-  );
-  const HipTotal = Object.values(HipCheckValue).reduce(
-    (sum, value) => sum + value,
-    0,
-  );
-  const JazzTotal = Object.values(JazzCheckValue).reduce(
-    (sum, value) => sum + value,
-    0,
-  );
-  const RbTotal = Object.values(RbCheckValue).reduce(
-    (sum, value) => sum + value,
-    0,
-  );
-  const RockTotal = Object.values(RockCheckValue).reduce(
-    (sum, value) => sum + value,
-    0,
-  );
-
-  const AllTotal = {
-    Pop: PopTotal,
-    Hip: HipTotal,
-    Jazz: JazzTotal,
-    Rb: RbTotal,
-    Rock: RockTotal,
-  };
-
-  const favoriteGenre = Object.keys(AllTotal).reduce((max, genre) => {
-    return AllTotal[genre] > AllTotal[max] ? genre : max;
-  }, 'Pop');
-
-  // 최대값
-  const maxTotal = AllTotal[favoriteGenre];
+  const favoriteGenre = useFavoriteGenre();
 
   return (
     <ScrollToNext>
